requirements.txt
requests~=2.25.0
python-dateutil~=2.8.1
beautifulsoup4~=4.9.3
databases~=0.4.0